+++ src/main.tsx
@@ -4,9 +4,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/App/App';
-import { registerServiceWorker } from './lib/registerServiceWorker';
 import './styles/global.scss';
 import './config/langConfig';
+import './lib/loadAssets';
 
 // Create root
 ReactDOM.createRoot(document.getElementById('root') as ReactDOM.Container).render(
@@ -14,8 +14,3 @@ ReactDOM.createRoot(document.getElementById('root') as ReactDOM.Container).rende
     <App />
   </React.StrictMode>,
 );
-
-// Request registration of SW
-window.addEventListener('load', () => {
-  registerServiceWorker();
-});