+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 ### Changed
 
 - Refactor App component
+- App icon
+
+### Removed
+
+- Unavailable Service Worker over HTTP
 
 ## [1.0.0] - 2026-07-13